Disable store module, add updateInteractions command
