+++ package.json
@@ -43,6 +43,7 @@
     "scryfall-sdk": "^3.1.0"
   },
   "devDependencies": {
+    "@oclif/core": "^1.16.0",
     "@types/chai": "^4.3.3",
     "@types/flexsearch": "^0.7.3",
     "@types/jsdom": "^20.0.0",
